FEATURE: Add new EnsureConsistency scheduled job
Add a new job that is scheduled twice daily to ensure the consistency of the calendar topic.

In some rare cases, like when a post is moved, the calendar is not properly updated.

Since reacting to a post being moved is hard (we don't have the information of the previous topics),
this job will ensure that the events of each posts in the calendar topic are up to date.