FIX: clean up holidays from deactivated/suspended/silenced users
Since we're generating holidays for the next 6 months, there might be some
leftovers whenever a user is deactivated/suspended/silenced.

PERF: only load timezones informations from concerned users.
We only care about the timezones of the users we will be generating holidays for.

DEV: rename user_ids_and_timezones to the clear-but-shorter timezones