FIX: Cannot compare nil with less than or equal to (#97)
This commit is to fix this error:

`Job exception: undefined method `<=' for nil:NilClass`

That occurs from the finish method in the monitor_events_dates scheduled
job.

```
app/models/discourse_post_event/event_date.rb:49:in `ended?'
jobs/scheduled/monitor_event_dates.rb:35:in `finish'
jobs/scheduled/monitor_event_dates.rb:11:in `block in execute'
```

This method checks to see if the event has finished and if it has it
will mark it as finished in the database. Some events though don't have
an end date and so this job will throw the above error.

Added a spec file for the event_date model to test for this specific
case.