FIX: Independence Day should use to_weekday_if_weekend(date) for all US
When Independence Day does not fall on a weekend, it is observed on a weekday. This was only happening for the state of Virginia. It has been updated to conform with the way things are handled for holidays like New Years Day and Christmas.

See https://publicholidays.us/independence-day/ which is also referenced on line 66.