FIX: do not hide upcoming events calendar when navigating away from page
Previously the ID for standard upcoming events calendar and category
events calendar was same and on every page change we were hiding the
category events calendar. Due to this the upcoming events calendar were
getting hidden when navigating away from page but was not reloading on
visiting again. This commit changes the ID for category events calendar
so that we can have finer control over it.

Internal ticket t61598.