DEV: Add API endpoint for holidays
This will add a new endpoint to return all of the holidays (as defined
by the Holidays gem) for a given region.

The new endpoint is:
`/admin/discourse-calendar/holiday-regions/:id/holidays`

This endpoint will only be available for admins and if the calendar
plugin is enabled with: `SiteSetting.calendar_enabled`.

If an invalid region is provided it will return a status code 422 with
an error message.