UX: Add admin holidays list view
This will add a new "Calendar" sidebar item in the admin area under
"Plugins". This new page will allow admins to select a region and then
all of the holidays for that region will be displayed.

Another notable change is that this will make `false` the default for
`allowNoneRegion` so it will only display calendar regions supported by
the Holidays gem by default.