FIX: delete event posts not automatic holidays
Renamed the 'DestroyPastEvents' job to 'DeleteExpiredPostEvents' and changed its behavior
to only delete posts (and their replies) that have an expired event.

Previously it would also delete expired 'CalendarEvents' but that was messing with our automated
holidays tracking.

Used 'post.reply_ids' to retrieve all the replies to a post, and not only the direct replies.
This helps keep the topic cleaner by removing the whole reply chain.