DEV: Add support for the experimental user menu (#313)
This commit ports the widgets-based implementation of the `event_reminder` and `event_invitation` notification items to the experimental Glimmer-based user menu. The changes in this commit should make event notifications in the experimental user menu consistent with they look/behave in the old user menu.

We don't need to add a `.discourse-compatibility` entry for this change because it's fully backward compatible since we check that the `registerNotificationTypeRenderer` plugin API is available before we attempt to use it.

For more context on the experimental user menu, see https://github.com/discourse/discourse/pull/17379.