FEATURE: handle subsequent and intersected holidays (#342)
If a user has several holidays one after another, we want to show the farthest end date.

For example, let's say today is Monday, and I am sick, and I also have days off from Tuesday to Friday. Then we want to show Friday as an end date of my time off.

Or let's say I have a vacation from Monday to Friday and at the same time there is a public holiday in my country on Monday. We want to show Friday as the end of my time off in this case.