FEATURE: closed events
Adds support for "closed" events instead of just relying on the start & end dates.

An event might now be "expired" when it happened in the past.

An event might now also be "closed", if for whatever reasons, the author decides to "cancel" it.

Context - https://meta.discourse.org/t/when-closing-event-it-moves-it-to-todays-date-time/307292