FEATURE: allow event editors to control list of users on the event
Previously event editors could remove people from an event but had no way
of acting on behalf of users in the event and adding them.

That meant that for events to properly show up in agenda and so on a user
must actively click a button.

In some cases (company ran events) the event manager may prefer controlling
attendance.