FIX: correctly apply local dates on event dates (#671)
The next ensures `this.htmlDates` has correctly been set and a render is on going which is going to be awaited by the schedule render. Before this fix `querySelectorAll` could return an empty nodes list as the dates were not rendered yet.

Note next shouldn't have this effect here, so it's either a false positive or another side effect we were not expecting.