FEATURE: Add who can see invitees setting
For some forums it might be useful to restrict who can see the list of invitees for an event without having to give `acting`/editing powers to a user.

This commit adds a new site setting `who_can_see_invitees` that allows admins to restrict who can see the list of invitees for an event.

The setting can be set to one of the following values:
- `only_who_can_act`: (default/current) only users who can act on the event can see the list of invitees.
- `all_logged_in_users` : all logged in users can see the list of invitees.
- `event_creator_and_who_can_act`: only the event creator and users who can act on the event can see the list of invitees.

- [ ] Add tests