FIX: remove double quotes `"` character when building the event's mar… (#716)
This is not the best UX as user might enter an event name with double quotes and they'll be deleted once they click save. But at least it won't break their event because the double quote breaks the BBCode/Markdown.

A proper fix would be to manipulate an AST instead of using regular expressions on a a string.

Meta - https://meta.discourse.org/t/-/360010