FIX: ensures UpcomingEventsCalendar can be used alone
This is a recent regression due to https://github.com/discourse/discourse-calendar/commit/bdf8869a0186264148ad95a2659fd863400db9e3

I wanted to write a test, but this is non trivial given this is a standalone component. We also have other internal repos testing this behavior (how I discovered the regression). Will try to have  a test for this one later.