FIX: Destroy calendar popover on navigation (#761)
This fixes an issue where we would show the popover for
a calendar event, but then not destroy it when navigating
to another page, which meant it ended up stuck in the top
left of the page.

This brings the behavior in line with other places we do
`destroyPopover()`, like when the mouse leaves the event
element.